added files
I originally pair programmed this assignment with Dominic in class.  I have been trying to do it myself but am in an endless loop of issues with trying to commit/push these changes.  Given the time of night I just wanted to show that I attempted to do this one!